[Maintenance] Change from float32 to float64
All that additional casting complexity isn't worth it, and it actually
*was* too little precision for some of the test values I tested the
upcoming currency formatter with!

Part of P0017, funded by qp.

+++ db_view.go
@@ -57,7 +57,7 @@ func PushesDelivered() PushProjection {
 // get done.
 type DiffInfoWeighted struct {
 	DiffInfo
-	Pushes float32
+	Pushes float64
 }
 
 // DiffsForEstimate returns the diffs of all completed pushes that are part of
